+++ tests/test_model/test_base.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import peewee
+from playhouse.test_utils import count_queries
 
 
 async def test_base():
@@ -47,15 +48,29 @@ async def test_backref(test_model, manager, schema):
     test = await ref.test
     assert test == source
 
-    assert ref.test == source
-    assert await ref.test == source
+    # Check cache
+    with count_queries() as counter:
+        assert await ref.test == source
+        assert await ref.test == source
+
+    assert counter.count == 0
 
     ref = await Ref.select(Ref, BaseModel).join(BaseModel).first()
     assert ref
-    assert ref.test == source
+
+    # Check preload
+    with count_queries() as counter:
+        assert await ref.test == source
+
+    assert counter.count == 0
 
     ref = await Ref(data="ref", test=test).save()
-    assert ref.test == source
+
+    # Check initiated
+    with count_queries() as counter:
+        assert await ref.test == source
+
+    assert counter.count == 0
 
     await Ref.drop_table()
 
@@ -71,12 +86,13 @@ async def test_fk(test_model, manager, schema):
     await ParentModel.create_table()
 
     parent = await ParentModel.create()
-    assert parent.child is None
+    assert parent.child_id is None
+    assert await parent.child is None
 
     child = await test_model.create(data="body")
     parent = await ParentModel.create(child=child)
 
-    assert parent.child == child
+    assert await parent.child == child
 
     await ParentModel.drop_table()
 
@@ -99,7 +115,7 @@ async def test_deferred_fk(manager):
 
     parent = ParentModel(child=child.id)
     assert await parent.child == child
-    assert parent.child == child
+    assert await parent.child == child
 
     await ParentModel.drop_table()
     await ChildModel.drop_table()